backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic-settings==2.7.1
email-validator==2.2.0
pymongo==4.10.1
certifi==2025.1.31
